split groups and invites sagas into their own separate dirs and change names to provide standard and to better reflect what they do
