+++ src/login/redux/sagas/groups/getAllDataSaga.js
@@ -2,7 +2,7 @@ import { call, put } from "redux-saga/effects";
 import { getData } from "../getDataRequest";
 import * as actions from "../../actions/getAllGroupMgmtDataActions";
 
-export function* allDataSaga() {
+export function* getAllDataSaga() {
   const url = GROUP_MGMT_URL + "/all-data";
   try {
     console.log("this is url", url);
