+++ src/login/redux/sagas/invites/getAllOutgoingInvitesSaga.js
@@ -2,7 +2,7 @@ import { call, put } from "redux-saga/effects";
 import { getData } from "../getDataRequest";
 import * as actions from "../../actions/getOutgoingInvitesActions";
 
-export function* allOutgoingInvitesSaga() {
+export function* getAllOutgoingInvitesSaga() {
   const url = GROUP_MGMT_URL + "/invites/outgoing";
   try {
     const allOutgoingInvitesResponse = yield call(getData, url);
